refactor: optimize folder structure && code refactor
