+++ src/components/Layout.tsx
@@ -3,8 +3,8 @@ import { Outlet } from 'react-router-dom';
 import { useAppSelector } from 'hooks/typedHooks';
 import { notificationSelector } from 'store/notificationSlice';
 
-import Footer from './layout/footer/Footer';
-import Header from './layout/header/Header';
+import Footer from './footer/Footer';
+import Header from './header/Header';
 import Loader from './Loader';
 import ToastBox from './ToastBox';
