+++ src/constants/constants.ts
@@ -17,8 +17,6 @@ export enum ApiRoutes {
   points = '/points',
 }
 
-export const _ = undefined; // for skipping unnecessary fn params
-
 export const TOKEN = 'token';
 
 export const LANG_TYPE = 'langType';
