+++ src/index.scss
@@ -1,3 +1,5 @@
+@import url(https://fonts.googleapis.com/css?family=Roboto:100,100italic,300,300italic,regular,italic,500,500italic,700,700italic,900,900italic);
+
 body {
   margin: 0;
   padding: 0;
@@ -5,7 +7,7 @@ body {
     'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  background:rgb(0, 53, 44);
+  background: rgb(0, 53, 44);
   scroll-behavior: smooth;
 }
 
@@ -43,7 +45,7 @@ input:-webkit-autofill:active {
 
 /* scrollbar Handle on hover */
 ::-webkit-scrollbar-thumb:hover {
-  background:rgb(0, 77, 64);
+  background: rgb(0, 77, 64);
 }
 
 /* Board description scrollbar */
@@ -56,6 +58,6 @@ input:-webkit-autofill:active {
     border-radius: 3px;
   }
   &::-webkit-scrollbar-thumb:hover {
-    background: rgb(161, 161, 173)
+    background: rgb(161, 161, 173);
   }
 }
