+++ src/index.tsx
@@ -42,7 +42,6 @@ const container = document.getElementById('root')!;
 const root = createRoot(container);
 
 root.render(
-  // <React.StrictMode>
   <Provider store={store}>
     <ThemeProvider theme={theme}>
       <BrowserRouter>
@@ -51,5 +50,4 @@ root.render(
       </BrowserRouter>
     </ThemeProvider>
   </Provider>
-  // </React.StrictMode>
 );
